#121 feat : 최근 접근한 가계부 키 저장
