+++ src/main/java/com/floney/floney/book/dto/process/DatesDuration.java
@@ -1,0 +1,29 @@
+package com.floney.floney.book.dto.process;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+import java.time.LocalDate;
+
+@RequiredArgsConstructor
+@Getter
+public class DatesDuration {
+    private LocalDate startDate;
+
+    private LocalDate endDate;
+
+    @Builder
+    private DatesDuration(LocalDate startDate, LocalDate endDate) {
+        this.startDate = startDate;
+        this.endDate = endDate;
+    }
+
+    public LocalDate start(){
+        return startDate;
+    }
+
+    public LocalDate end(){
+        return endDate;
+    }
+}
