+++ src/main/java/com/floney/floney/book/dto/request/SaveRecentBookKeyRequest.java
@@ -1,0 +1,8 @@
+package com.floney.floney.book.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class SaveRecentBookKeyRequest {
+    private String bookKey;
+}
